web-animations/animation-sequence/animation-sequence-synchronization.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationSequence child nodes synchronization test</title>
<meta name="assert" content="The start time of a child animation node of an animation sequence is the end time of the child's previous sibling. If the child has no previous sibling the start time is zero.">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#the-start-time-of-children-of-an-animation-sequence">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var animationSequence = new AnimationSequence([node1, node2]);
    var player = document.timeline.play(animationSequence);
    animationSequence.player.currentTime = ANIMATION_END_TIME/2;

    assert_equals(window.getComputedStyle(node1.target).width, 100,
        'Animation should take place for the first node in the sequence');
    assert_equals(window.getComputedStyle(node2.target).width, 10,
            'Animation should not take place for the second node in the sequence');
}, 'Nodes in the hierarchy are controlled by parent group player');
</script>
</body>
